fix(ci): refresh genie fod hash
